requirements.txt
mlflow==2.17.0
cloudpickle==2.2.1
nltk==3.9.1
numpy==1.23.5
psutil==5.9.0
scikit-learn==1.4.2
scipy==1.11.4
Flask
azure-storage-blob
gunicorn
deep-translator
# suivi dans Azure Application Insight
#azure-monitor-opentelemetry-exporter
#opentelemetry-api
#opentelemetry-sdk
opencensus-ext-azure